fixed commands classes part/name for rails 7 naming conventions
